Move constant to more appropriate file
