--- parse_this/values.py
@@ -1,1 +1,0 @@
-_NO_DEFAULT = object()
